feat: Move name_hash into db package
