+++ backend/cmd/hub/key_rotate.go
@@ -12,7 +12,6 @@ import (
 
 	"charm.land/lipgloss/v2"
 	"github.com/OrcaCD/orca-cd/internal/hub"
-	"github.com/OrcaCD/orca-cd/internal/hub/applications"
 	"github.com/OrcaCD/orca-cd/internal/hub/crypto"
 	"github.com/OrcaCD/orca-cd/internal/hub/db"
 	"github.com/OrcaCD/orca-cd/internal/hub/models"
@@ -145,7 +144,7 @@ func runKeyRotateCommandWithInput(ctx context.Context, out io.Writer, in io.Read
 
 	// The application name_hash blind index is keyed off APP_SECRET, so rotation
 	// invalidates every stored hash. Recompute them under the new key.
-	if err := applications.BackfillNameHashes(ctx); err != nil {
+	if err := db.BackfillNameHashes(ctx, db.DB); err != nil {
 		return fmt.Errorf(
 			"key rotation re-encrypted the database but failed to recompute application name hashes: %w. Restart the hub with the new APP_SECRET to retry the backfill",
 			err,
