+++ backend/internal/hub/db/name_hash.go
@@ -1,10 +1,9 @@
-package applications
+package db
 
 import (
 	"context"
 
 	"github.com/OrcaCD/orca-cd/internal/hub/crypto"
-	"github.com/OrcaCD/orca-cd/internal/hub/db"
 	"github.com/OrcaCD/orca-cd/internal/hub/models"
 	"gorm.io/gorm"
 )
@@ -13,8 +12,8 @@ import (
 // using the current default cipher. It is idempotent and is run on startup (to
 // populate rows created before the column existed) and after key rotation (when
 // the blind-index key derived from APP_SECRET has changed).
-func BackfillNameHashes(ctx context.Context) error {
-	apps, err := gorm.G[models.Application](db.DB).Find(ctx)
+func BackfillNameHashes(ctx context.Context, gdb *gorm.DB) error {
+	apps, err := gorm.G[models.Application](gdb).Find(ctx)
 	if err != nil {
 		return err
 	}
@@ -24,7 +23,7 @@ func BackfillNameHashes(ctx context.Context) error {
 		if hash == apps[i].NameHash {
 			continue
 		}
-		if _, err := gorm.G[models.Application](db.DB).
+		if _, err := gorm.G[models.Application](gdb).
 			Where("id = ?", apps[i].Id).
 			Select("NameHash").
 			Updates(ctx, models.Application{NameHash: hash}); err != nil {
